Change Go module to github.com/DataDog/rules_oci_go
This is so that Gazelle can actually import the repository as a Go repository when we already have `@com_github_datadog_rules_oci` used for the rules themselves. Gazelle will try to generate the Go repository name based on the Go module name, so we need the Go module name to be distinct from the rules name.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>