feat(oci_image): adds `env` attribute to `oci_image`
According to the OCI image spec, images (but not layers) can also specify
environment variables. This adds that capability to images created from
this ruleset.

I also added a somewhat hacky bootstrap target that just copies the OCI
tool binary into the four possible locations it should go. Unfortunately
we can't use a `select` and `write_source_file` because the `out_file`
attribute is not passed directly to a rule (it gets used in the macro).