chore: update go and rules_go version (#55)
older versions of Go suffer from issues with DNS resolution when
cross-compiling. the root of the issue is that Go <1.20 uses a built-in
DNS resolver which doesn't know how to resolve addresses when connected
to any VPN, like AppGate. because of this, rules like `oci_push` cannot
push images to GovCloud registry, making it more difficult to do things
like deploy from a laptop in the event of an incident.

this DNS issue was fixed in 1.20 (see [this
article](https://danp.net/posts/macos-dns-change-in-go-1-20/)) for more
information.